// src/app/interfaces/medidas.ts
export interface Medidas {
  idmedidas?: number;
  idmatricula?: number;
  fecha?: string;
  talla?: number;
  peso?: number;
  edad?: number;
  busto_alto?: number;
  busto_bajo?: number;
  cintura?: number;
  cadera_abdomen?: number;
  cadera_gluteos?: number;
  brazo_derecho?: number;
  brazo_izquierdo?: number;
  pierna_derecha?: number;
  pierna_izquierda?: number;
  guia_nutricional?: number;
  vigencia?: number;
}